refactor(bpd): decompor BrazilPromotionDayLanding em 6 subcomponentes
